requirements.txt
streamlit>=1.36
python-dotenv>=1.0
langchain>=0.3
langchain-ollama>=0.0.6
ollama>=0.3.0
pydantic>=2.8
